langchaing/requirements.txt
langchain-community==0.2.12
langchain-openai==0.1.22
python-dotenv==1.0.1